Update seeds to not have all positions that have ended in the past
Not sure if it was by design, but all positions created in the seed file will have ended in the past.

This fixes it, but not sure if having an end date in the future is better. Maybe having a local variable that holds a random number in the range `(-50..50)` then setting `ended_at` to nil if the number is negative.

Let me know if you have any questions about it.